+++ CHANGELOG.md
@@ -8,11 +8,16 @@ All notable changes to the "remvscode" extension will be documented in this file
 
 - Menus in Header to select what is being refactored
 - Preview Window to review changes
+- Inbuilt (Read-only) Editor to see the code that has been selected
+    - If the user only selects part of a line, the whole line will be selected
+      automatically
 - Confirm Button to set changes (at this stage, no changes are being applied)
 
 ### Changed
+- None
 
 ### Removed
+- None
 
 ## [1.0.0]